refactor(src) remove all
