+++ src/api/auth/auth.py
@@ -1,7 +1,7 @@
 import uuid
 from fastapi_users.authentication import CookieTransport, AuthenticationBackend
 from fastapi_users.authentication import JWTStrategy
-from app.auth.core import get_user_manager
+from api.auth.core import get_user_manager
 from database.engine import settings
 from fastapi_users import FastAPIUsers
 from models.users import UserORM
