+++ src/app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
 from .config import Settings
-from routers import include_routers
+from api.dependencies import include_routers
 
 app = FastAPI(**Settings().app_presets)
 include_routers(app)
