+++ src/database/engine/pgsql.py
@@ -1,5 +1,5 @@
-from .config import Settings
-from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker
+from app.config import Settings
+from sqlalchemy.ext.asyncio import create_async_engine
 
 settings = Settings()
 pgs_url = f"{settings.DB_USER}:{settings.DB_PASS}@{settings.DB_HOST}:{settings.DB_PORT}/{settings.DB_NAME}"
