+++ src/services/dynamic.py
@@ -2,11 +2,10 @@ import json
 from typing import Annotated
 from uuid import UUID
 from datetime import datetime
-from fastapi import Query
 from pydantic import create_model, BaseModel, ValidationError, Field
 from database.repositories import CamerusRepo
 from typing import Type
-from database.models import DependenciesOrm
+from models import DependenciesOrm
 
 
 class DynamicModels:
